Rename knight client api and service
